sigma-ai-main/requirements.txt
flask
flask-cors
google-generativeai
sympy
gunicorn
